algotrader3/requirements.txt
# AlgoTrader3 World Model Requirements
# Install with: pip install -r requirements.txt

# Core ML/Deep Learning
torch>=2.0.0
numpy>=1.24.0

# Reinforcement Learning Environment
gymnasium>=0.29.0

# Data Processing
pandas>=2.0.0
pandas-ta>=0.3.14b  # Technical indicators

# Normalization/Scaling
scikit-learn>=1.3.0

# Evolutionary Optimization (CMA-ES for controller training)
cma>=3.3.0

# Configuration
pyyaml>=6.0

# Progress bars
tqdm>=4.65.0

# Visualization (optional, for evaluation plots)
matplotlib>=3.7.0

# Logging
colorlog>=6.7.0  # Optional: colored logging
